requirements.txt
six==1.10.0
django-import-export==0.5.1
django==1.11.12
mock==2.0.0
coverage>=4.4.1
celery==4.2.0
flower==0.9.3
pyyaml==5.2
kombu==4.2.1
python-memcached==1.59
djangorestframework==3.7.7
django-filter==1.1.0
elasticsearch==6.2.0
PyMySQL==0.9.2
natsort==5.3.3
ldap3==2.5.2
